+++ client/boardspace-java/boardspace-games/havannah/HavannahViewer.java
@@ -329,8 +329,8 @@ public class HavannahViewer extends CCanvas<HavannahCell,HavannahBoard> implemen
     	CELLSIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*CELLSIZE);
-    	int boardH = (int)(nrows*CELLSIZE);
+    	int boardW = (ncols*CELLSIZE);
+    	int boardH = (nrows*CELLSIZE);
     	int extraW = Math.max(0, (mainW-boardW)/2);
     	int extraH = Math.max(0, (mainH-boardH)/2);
     	int boardX = mainX+extraW;
@@ -412,7 +412,7 @@ public class HavannahViewer extends CCanvas<HavannahCell,HavannahBoard> implemen
             while (nc-- > 0)
             {	int rx = Random.nextInt(rand, spacex);
                 int ry = Random.nextInt(rand, spacey);
-                c.drawChip(gc,this,chip.getAltDisplayChip(rx+ry),(int)cs,G.Left(r)+cs/2+rx,G.Top(r)+cs/2+ry,null);
+                c.drawChip(gc,this,chip.getAltDisplayChip(rx+ry),cs,G.Left(r)+cs/2+rx,G.Top(r)+cs/2+ry,null);
              }
         }
     }
