+++ client/boardspace-java/boardspace-games/quinamid/QuinamidViewer.java
@@ -234,8 +234,8 @@ public class QuinamidViewer extends CCanvas<QuinamidCell,QuinamidBoard> implemen
     	SQUARESIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*SQUARESIZE);
-    	int boardH = (int)(nrows*SQUARESIZE);
+    	int boardW = (ncols*SQUARESIZE);
+    	int boardH = (nrows*SQUARESIZE);
     	int extraW = Math.max(0, (mainW-boardW)/2);
     	int extraH = Math.max(0, (mainH-boardH)/2);
     	int boardX = mainX+extraW;
